Added a clause about state of website
This website is only for practicing HTML/CSS.  Depending on when it is accessed it could be beautiful or a complete mess.  It's a scratch pad I work on when I'm bored or just want to practice new techniques.  However, the information is valuable for those wanting to know a little more about me.